+++ app/services/things/record_visit.rb
@@ -1,0 +1,7 @@
+module Things
+  class RecordVisit
+    def self.call(thing:)
+      thing.increment!(:visit_count)
+    end
+  end
+end